+++ android/app/src/main/kotlin/dev/xsoulspace/lastanswer/MainActivity.kt
@@ -2,5 +2,4 @@ package dev.xsoulspace.lastanswer
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity : FlutterActivity()