+++ packages/core/lib/src/data_sources/local_hive/hive_models.dart
@@ -5,7 +5,7 @@ import 'dart:convert';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/widgets.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:hive/hive.dart';
+import 'package:hive_ce/hive.dart';
 import 'package:mockito/mockito.dart';
 import 'package:shared_models/shared_models.dart';